backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
pypdf==4.0.2
reportlab==4.1.0
Pillow==10.2.0
python-dotenv==1.0.1
aiofiles==23.2.1
slowapi==0.1.9
redis==5.0.1
apscheduler==3.10.4
structlog==24.1.0
httpx==0.27.0
email-validator==2.1.0.post1